requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
channels==4.0.0
channels-redis==4.1.0
asgiref==3.7.0
requests==2.31.0
twilio==8.10.0
